pyproject.toml
[project]
name = "syntsch"
version = "0.1.0"
description = "Autonomous AI-powered cultural digest for Berlin. Finds events, writes editorial essays, publishes daily."
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0",
]

[tool.setuptools]
py-modules = ["cli", "config", "models", "storage"]
packages = ["agents", "sources", "notifiers"]

[project.urls]
Homepage = "https://syntsch.de"
Repository = "https://github.com/maksugr/syntsch"

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "pytest>=8.0",
]
